Added logging and fixed dependency issues
Added logging to check in views and checkin.js file to help debug the check in feature. I fixed dependency issues within docker by getting rid of mui-phone-number and replace it with mui-tel-input which supports MUI V6. I also made sure MUIV6 was only version of mui we were using. Tested in docker and seems to be working as of this push.